+++ django_stripe/admin/__init__.py
@@ -1,4 +1,13 @@
 from django_stripe.admin.core import StripeCustomerAdmin
-from django_stripe.admin.products import StripeProductAdmin, StripeCouponAdmin
+from django_stripe.admin.products import (
+    StripeProductAdmin,
+    StripePriceAdmin,
+    StripeCouponAdmin,
+)
 
-__all__ = ("StripeCustomerAdmin", "StripeProductAdmin")
+__all__ = (
+    "StripeCustomerAdmin",
+    "StripeProductAdmin",
+    "StripePriceAdmin",
+    "StripeCouponAdmin",
+)