+++ docs/changelog.md
@@ -14,10 +14,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `StripeCardAction` action that will sync card data from stripe.
 
 
+## [0.2.0] - 2024-09-16
+-------------------
+### Added
+* `StripeCustomerAdmin` admin class that will display stripe customer data.
+* `StripeEventAdmin` admin class that will display stripe event data.
+* `StripeProductAdmin` admin class that will display stripe product data.
+* `StripePriceAdmin` admin class that will display stripe price data.
+* `StripeCouponAdmin` admin class that will display stripe coupons data.
+* `StripeSubscriptionAdmin` admin class that will display stripe subscription data.
+
+### Changed
+* renamed `date_purged` to `deleted_at` for stripe models.
+
+### Fixed
+* Sync for `StripePriceAction` failing.
+
+
 ## [0.1.1] - 2024-09-16
 -----------------------
 ### Added
-
   - **Stripe Models:** The following models have been added to simplify
     the integration with Stripe.
     - `StripeCustomer` model that will store stripe customer data.